Move torrent magnet and torrent download into 'torrents' blueprint
