+++ nyaa/routes.py
@@ -1,9 +1,6 @@
-import os.path
-from urllib.parse import quote
-
 import flask
 
-from nyaa import api_handler, app, db, forms, models, template_utils, torrents, views
+from nyaa import api_handler, app, db, forms, models, template_utils, views
 from nyaa.backend import get_category_id_map
 
 DEBUG_API = False
@@ -48,32 +45,6 @@ def delete_comment(torrent_id, comment_id):
     return flask.redirect(url)
 
 
-@app.route('/view/<int:torrent_id>/magnet')
-def redirect_magnet(torrent_id):
-    torrent = models.Torrent.by_id(torrent_id)
-
-    if not torrent:
-        flask.abort(404)
-
-    return flask.redirect(torrents.create_magnet(torrent))
-
-
-@app.route('/view/<int:torrent_id>/torrent')
-@app.route('/download/<int:torrent_id>.torrent')
-def download_torrent(torrent_id):
-    torrent = models.Torrent.by_id(torrent_id)
-
-    if not torrent or not torrent.has_torrent:
-        flask.abort(404)
-
-    resp = flask.Response(_get_cached_torrent_file(torrent))
-    resp.headers['Content-Type'] = 'application/x-bittorrent'
-    resp.headers['Content-Disposition'] = 'inline; filename="{0}"; filename*=UTF-8\'\'{0}'.format(
-        quote(torrent.torrent_name.encode('utf-8')))
-
-    return resp
-
-
 @app.route('/view/<int:torrent_id>/submit_report', methods=['POST'])
 def submit_report(torrent_id):
     if not flask.g.user:
@@ -94,17 +65,6 @@ def submit_report(torrent_id):
         flask.flash('Successfully reported torrent!', 'success')
 
     return flask.redirect(flask.url_for('torrents.view', torrent_id=torrent_id))
-
-
-def _get_cached_torrent_file(torrent):
-    # Note: obviously temporary
-    cached_torrent = os.path.join(app.config['BASE_DIR'],
-                                  'torrent_cache', str(torrent.id) + '.torrent')
-    if not os.path.exists(cached_torrent):
-        with open(cached_torrent, 'wb') as out_file:
-            out_file.write(torrents.create_bencoded_torrent(torrent))
-
-    return open(cached_torrent, 'rb')
 
 
 # #################################### BLUEPRINTS ####################################
